fix error float

+++ Sources/svglib/circle.cpp
@@ -16,6 +16,6 @@ string Circle::generate(float x, float y, float r) {
     fill.assign(properties.fill.fill);
     string stroke;
     stroke.assign(properties.stroke.stroke);
-    string s = "<circle cx=\"" + to_string((int) x) + "\" cy=\"" + to_string((int) y) + "\" fill=\"" + fill + "\" r=\""+ to_string((int) r) +"\" stroke=\""+ stroke +"\" stroke-width=\""+ to_string((int) properties.stroke.stroke_width) +"\"/>";
+    string s = "<circle cx=\"" + to_string(x) + "\" cy=\"" + to_string(y) + "\" fill=\"" + fill + "\" r=\""+ to_string(r) +"\" stroke=\""+ stroke +"\" stroke-width=\""+ to_string((int) properties.stroke.stroke_width) +"\"/>";
     return s;
 }
